+++ routes/web.php
@@ -36,7 +36,7 @@ Route::get('/register/vendor', function () {
 })->name('register')->middleware('guest');
 Route::get('/vendorList',[VendorController::class, 'index']);
 Route::get('/orderList/{id}',[OrderController::class, 'viewOrderList']);
-Route::get('/checkout',[CartController::class, 'cartIndex']);
+Route::get('/checkout',[ProductController::class, 'cartIndex']);
 Route::get('/register/{lang}', function ($lang) {
     App::setLocale($lang);
     return view ('register');
